Run the `rest2yaml` sub-command added in the previous commit.
This commit is a result of running the following command:

    $ python scripts/vip.py rest2yaml

It parses the reST files, generates YAML files from the data
in the reST tables, and rewrites the reST files to use
the tables generated from the YAML files.

The `update_tables` sub-command needs to be run subsequently
in order for the new tables to be present.