Increase precision of LatLng elements.
Lat/long coordinates frequently are given with greater precision than can
accurately be represented in a single-precision (32-bit) float. For example,
the lat/long pair "43.038217, -77.711050" when parsed as a pair of 32-bit
floats, produces a pair of numbers (to six significant figures) 43.038214,
-77.711053.

While the error introduced by round-tripping a lat-long pair in the US through
a float should usually amount to well under a meter of actual distance (in the
given example it's off by about 0.3m), the cost of using
double-precision (64-bit) floats, which can accurately represent a lat/long
pair in degrees to about a nanometer, is not meaningful, and the format change
is fully forwards-compatible (since the string representations of float and
double are identical).