Create place_id.yaml
A PlaceId is a way to refer to a location based on a geocoding or other data provider.  The supported providers are outside the scope of the standard, because they are likely to change; users are free to ignore place IDs from providers they do not support.

Examples uses of this entity include Google Maps Place Ids, Google Plus Codes, or Latitude/Longitude from a specific geocoder.